Fix linux prepare

+++ src/gumbo/error.cpp
@@ -80,7 +80,7 @@ static void print_tag_stack(
     if (i) {
       print_message(parser, output, ", ");
     }
-    GumboTag tag = (GumboTag)(unsigned int) error->tag_stack.data[i];
+    GumboTag tag = (GumboTag)(uintptr_t) error->tag_stack.data[i];
     print_message(parser, output, gumbo_normalized_tagname(tag));
   }
   gumbo_string_buffer_append_codepoint(parser, '.', output);
